requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
PyJWT[crypto]>=2.8.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
